refactor(api): consolidar rutas de API de usuarios y adoptantes
- Fusionar /api/user/adoptions y /api/adopter/adoptions en /api/adoptions
- Mover /api/adopter/favorites a /api/user/favorites
- Actualizar /api/users/profile a /api/user/profile
- Actualizar las llamadas de obtención de datos relacionadas para usar las nuevas rutas de API
